requirements.txt
python-telegram-bot
python-dotenv
sqlalchemy
psycopg2-binary
alembic